// js/src/types/resources/ClinicalImpression.ts
 
 
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TDateTime } from '../simpleTypes/DateTime';
import { TPeriod } from '../partials/Period';
import { TClinicalImpressionInvestigation } from '../partials/ClinicalImpressionInvestigation';
import { TClinicalImpressionFinding } from '../partials/ClinicalImpressionFinding';
import { TAnnotation } from '../partials/Annotation';

export type TClinicalImpression = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  status: string;
  statusReason?: TCodeableConcept;
  code?: TCodeableConcept;
  description?: string;
  subject: TReference;
  encounter?: TReference;
  effectiveDateTime?: TDateTime;
  effectivePeriod?: TPeriod;
  date?: TDateTime;
  assessor?: TReference;
  previous?: TReference;
  problem?: TReference[];
  investigation?: TClinicalImpressionInvestigation[];
  protocol?: TUri[];
  summary?: string;
  finding?: TClinicalImpressionFinding[];
  prognosisCodeableConcept?: TCodeableConcept[];
  prognosisReference?: TReference[];
  supportingInfo?: TReference[];
  note?: TAnnotation[];
};